src/test/java/it/mds/sdk/flusso/dispovig/regole/custom/RegolaUguaglianzaCrossCampoCondizionatoTest.java
/* SPDX-License-Identifier: BSD-3-Clause */

package it.mds.sdk.flusso.dispovig.regole.custom;

import it.mds.sdk.flusso.dispovig.regole.custom.adhoc.RegolaDispovigCodStruttSan;
import it.mds.sdk.gestoreesiti.modelli.Esito;
import it.mds.sdk.libreriaregole.dtos.RecordDtoGenerico;
import it.mds.sdk.libreriaregole.exception.ValidazioneImpossibileException;
import it.mds.sdk.libreriaregole.regole.beans.Campo;
import it.mds.sdk.libreriaregole.regole.beans.Parametri;
import it.mds.sdk.libreriaregole.regole.beans.RegolaGenerica;
import it.mds.sdk.libreriaregole.regole.beans.RegoleFlusso;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.Mock;
import org.mockito.Mockito;
import org.mockito.MockitoAnnotations;
import org.mockito.junit.jupiter.MockitoExtension;

import java.lang.reflect.InvocationTargetException;
import java.util.ArrayList;
import java.util.HashMap;
import java.util.List;
import java.util.Map;

import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.spy;

@ExtendWith(MockitoExtension.class)
class RegolaUguaglianzaCrossCampoCondizionatoTest {
    @Mock
    RecordDtoGenerico recordMockito;
    RegolaUguaglianzaCrossCampoCondizionato regola;
    Parametri parametriTest;
    @BeforeEach
    void init(){
        MockitoAnnotations.openMocks(this);
        Map<String, String> parametri = new HashMap<>();
        parametriTest = new Parametri();
        parametriTest.setParametriMap(parametri);

        regola = spy(new RegolaUguaglianzaCrossCampoCondizionato("TEST", "TEST", "TEST", parametriTest));
    }

    @Test
    void validaCostruttore() {
        RegolaUguaglianzaCrossCampoCondizionato regola = new RegolaUguaglianzaCrossCampoCondizionato();
        assertTrue(regola instanceof RegolaUguaglianzaCrossCampoCondizionato);
    }

    @Test
    void validaKO() throws InvocationTargetException, IllegalAccessException, NoSuchMethodException {
        Map<String, String> parametri = new HashMap<String, String>();

        parametri.put("campoValidatore", "campo1");
        parametri.put("parametroInput", "val1");
        parametri.put("campoUguaglianza", "campo2");

        Parametri parametriTest = new Parametri();
        parametriTest.setParametriMap(parametri);

        RegolaUguaglianzaCrossCampoCondizionato regola = new RegolaUguaglianzaCrossCampoCondizionato("RegolaUguaglianzaCrossCampoCondizionato", "9999", "descrizioneErrore",parametriTest);
        Mockito.when(recordMockito.getCampo("campoRecord")).thenReturn(null);
        Mockito.when(recordMockito.getCampo("campo1")).thenReturn("val1");
        Mockito.when(recordMockito.getCampo("campo2")).thenReturn("val2");

        List<Esito> result = regola.valida("campoRecord", recordMockito);
        for (Esito e : result) {
            assertFalse(e.isValoreEsito());
        }
    }
    @Test
    void validaOK() throws InvocationTargetException, IllegalAccessException, NoSuchMethodException {
        Map<String, String> parametri = new HashMap<String, String>();

        parametri.put("campoValidatore", "campo1");
        parametri.put("parametroInput", "val1");
        parametri.put("campoUguaglianza", "campo2");

        Parametri parametriTest = new Parametri();
        parametriTest.setParametriMap(parametri);

        RegolaUguaglianzaCrossCampoCondizionato regola = new RegolaUguaglianzaCrossCampoCondizionato("RegolaUguaglianzaCrossCampoCondizionato", "9999", "descrizioneErrore",parametriTest);
        Mockito.when(recordMockito.getCampo("campoRecord")).thenReturn("val1");
        Mockito.when(recordMockito.getCampo("campo1")).thenReturn("val1");
        Mockito.when(recordMockito.getCampo("campo2")).thenReturn("val1");

        List<Esito> result = regola.valida("campoRecord", recordMockito);
        for (Esito e : result) {
            assertTrue(e.isValoreEsito());
        }
    }
    @Test
    void validaKOException() throws IllegalAccessException, InvocationTargetException, NoSuchMethodException
    {
        Mockito.when(recordMockito.getCampo(any())).thenThrow(new IllegalAccessException());
        assertThrows(ValidazioneImpossibileException.class,()->regola.valida("codStruttura", recordMockito));
    }

}